+++ include/wx/wxsqlite3def.h
@@ -46,6 +46,45 @@
 
 <dl>
 
+<dt><b>4.10.12</b> - <i>August 2025</i></dt>
+<dd>
+Upgrade to <i>SQLite3 Multiple Ciphers version 2.2.4 (SQLite version 3.50.4)</i>
+
+</dd>
+
+<dt><b>4.10.11</b> - <i>July 2025</i></dt>
+<dd>
+Upgrade to <i>SQLite3 Multiple Ciphers version 2.2.3 (SQLite version 3.50.3)</i>
+to fix issue #127
+
+</dd>
+
+<dt><b>4.10.10</b> - <i>July 2025</i></dt>
+<dd>
+Upgrade to <i>SQLite3 Multiple Ciphers version 2.2.2 (SQLite version 3.50.2)</i>
+to fix issue #126
+
+</dd>
+
+<dt><b>4.10.9</b> - <i>July 2025</i></dt>
+<dd>
+Upgrade to <i>SQLite3 Multiple Ciphers version 2.2.1 (SQLite version 3.50.2)</i>
+
+</dd>
+
+<dt><b>4.10.8</b> - <i>June 2025</i></dt>
+<dd>
+Upgrade to <i>SQLite3 Multiple Ciphers version 2.1.3 (SQLite version 3.50.1)</i>
+
+</dd>
+
+<dt><b>4.10.7</b> - <i>June 2025</i></dt>
+<dd>
+Upgrade to <i>SQLite3 Multiple Ciphers version 2.1.2 (SQLite version 3.50.0)</i><br>
+Add method wxSQLite3Database::SetLockTimeout()
+
+</dd>
+
 <dt><b>4.10.6</b> - <i>May 2025</i></dt>
 <dd>
 Add method wxSQLite3Database::Configure()<br>